Done doing the total pay of employee for MVP

+++ src/main/java/com/art/timelymanagementsystem/dto/UserPayrollDto.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 @Data
 public class UserPayrollDto {
@@ -12,7 +13,7 @@ public class UserPayrollDto {
     private Long userId;
     private LocalDate startDate;
     private LocalDate endDate;
-    private BigInteger totalHours;
+    private BigDecimal totalHours;
     private BigDecimal totalPay;
     private BigDecimal hourlyRate;
